modularizacion de los componentes
