+++ src/App.jsx
@@ -2,8 +2,8 @@ import './App.css'
 import { Navbar } from './components/Navbar/Navbar';
 import { Promotions } from './components/Promotions/Promotions';
 import {Home} from './components/Home/Home';
-import {Cards} from './components/Servicios/Cards';
-import Ofertas from './components/Servicios/Ofertas';
+import {Cards} from './components/Cards/Cards';
+import Ofertas from './components/Ofertas/Ofertas';
 import Carrousel from './components/Carrousel/Carrousel';
 import Footer from '../src/components/Footer/Footer';
