+++ src/components/Card/Card.jsx
@@ -1,5 +1,6 @@
-
-import './Servicios.css';
+/* import './Servicios.css'; */
+import '../Cards/Servicios.css'; 
+i
 
 const Card = props => {
     return(
