+++ src/components/Cards/Cards.jsx
@@ -1,6 +1,6 @@
 
 import { useEffect,useState } from 'react';
-import Card from './Card'; 
+import Card from '../Card/Card'; 
 import axios from 'axios';
