+++ src/components/Modal/Modal_card/Modal_card.css
